Be more rigorous about sp_link refcounting etc.
Don't insert links more than once to the map.

m_trackLinkMap now owns a ref to each entry added to it, and releases it
whenever removeFromTrackLinkMap() is called, it releases this ref.
It doesn't really matter though, because removeFromTrackLinkMap() is never
called.

SpotifySearch:searchComplete() only holds a ref for long enough to convert
it to a uid.