Fixes a bug where the countryId is first set to '' and cannot be updated
By default the countryId is set to '' and further checks just verify
that the countryId exist, which returns true, but trying to perform a
search will fail due to the empty countryId. Also, the method that
retrieves the countryId uses the wrong context, and thus cannot
actually set the correct countryId.